+++ src/client/components/game-night-accordion/EventItem.jsx
