+++ src/client/components/game-night-accordion/NightDetails.jsx
