+++ src/client/components/game-night-form/DividedListItem.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types'
 import {
   ListItem,
   ListItemText,
-  Typography,
   Divider,
   IconButton,
 } from '@mui/material';
