+++ src/client/views/GameNights.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import axios from 'axios';
 import { Button, Typography } from '@mui/material';
 import Navbar from '../components/Navbar';
-import GameNightForm from '../components/GameNightForm';
+import GameNightForm from '../components/game-night-form/GameNightForm';
 import GameNightsList from '../components/GameNightsList';
 
 const {
